+++ activerecord/test/cases/migration/change_schema_test.rb
@@ -384,6 +384,17 @@ module ActiveRecord
         end
       end
 
+      def test_force_drop_nonexistent_table
+        assert_nothing_raised { connection.drop_table(:nonexistent, :force => true) }
+      end
+
+      def test_force_drop_existing_table
+        connection.create_table :testings
+        assert connection.table_exists?(:testings)
+        connection.drop_table(:testings, :force => true)
+        assert_not connection.table_exists?(:testings)
+      end
+
       private
       def testing_table_with_only_foo_attribute
         connection.create_table :testings, :id => false do |t|